Add remeining new VMs and wire for Avalonia UI
